First commit of richardson_lucy python class
This is the initial commit of a richardson_lucy python class.  In the
future this class has the potential to become quite complicated.  For
example there are different types of padding that can be in in
richardson lucy, different edge handling strategies, different types of
regularization, even different types of acceleration schemes.

For the initial commit we do the minimum to get it to run.  We need to
extend to the next smooth size so the low level code that is based on
clfft will work for all sizes.