Set OPENCL_INCLUDE_DIR for linux build
When testing on an Ubuntu 20.04 system I found that I needed to set
OPENCL_INCLUDE_DIR to get the build to work.  In this case the path is
set to point to the cuda include directory which contains the opencl
header files. On other systems and flavors of linux a different path may
be needed.