Handle non-circulant parameter properly
The boolean 'non-circulant' is at zero-based position 5, not 4.

TODO:  Review current design.  A child class can add more parameters
(for example so far we've added regularization factor and non-circulant)
which are parsed in the parent class, this means parent class would have
to be kept in sync with potential child class.  It only works if there
is only a small number of child classes.