Make sure libs are installed in 'lib'
We want to put the native libraries into the 'lib' directory under the
python environment directory so that they are found at runtime.

In anaconda the libraries end up being installed in

anaconda3\envs\napari-env\$env_name$\lib

and seem to be found when running in that environemnt.

TODO: Add apple and apple m1, test in non-conda virtual environments