Clean up debugging print statements
Replace most print statements with Macros called CHECK, CHECKRETURN and
ENTEREXIT.  Printing messages can then be turned off an on by both
individual flags for every print statement and by a global flag.

Also clean up the printing of progress messages for the iterative
updates.  Instead of printing a progress message on a unique line for
every iteration, print the iterations number, every 10th iterations and
keep the update to a single long line.  ie For a 100 iteration
deconvolution the message will look as below.

10 20 30 40 50 60 70 80 90 100