Add script to install in editable mode
This is needed because the native libraries don't seem to be copied to
the correct spot when installing in editable mode.  Thus the script
copies the native libraries to a location where they can be found in an
Annaconda or miniconda environment.

This script is meant to be run locally so the paths would need to be
changed on each machine to reflect the location of the
anaconda/miniconda installation and current environment.