Add print_info script
This calls the new print_platforms_and_devices function from the native
code which prints a list of platforms and devices available.  We now
allow the caller to specify which platform/device combination to run on,
so this little utility is handy to quickly check which platforms and
devices are available and there indexes.

For example on a system with an NVIDIA GPU and Intel processor the
output would be as below.

platform 0 NVIDIA CUDA
     device name 0 NVIDIA GeForce RTX 3090

platform 1 Intel(R) OpenCL
     device name 0 Intel(R) Core(TM) i9-10940X CPU @ 3.30GHz

Then the user could call the deconvolution with either platform 0,
device 0 (NVIDIA RTX 3090) or platform 1, device 0 (Intel i9)