Copy some tests into 'interactive tests'
Moving forward the 'tests' directory will be used for unit type tests
that will be run with pytest.  They will use dummy data to test if the
software is functioning right and producing correct values for simulated
cases.

'interactive tests' will be used to test and troubleshoot with real
files.