Work on progress and status updates.
The chunked deconvolutions can take a long time so it is nice for the
users to know what is going on.  We could use the imagej progress bar
and log window, but sometimes users don't notice those.  So I am
experimenting with putting a simple progress bar and log window right on
the GUI, then wrapping them in the scijava StatusService interface.